requirements.txt
Flask
Flask-Cors
python-dotenv
pymongo
scikit-learn
pandas
gunicorn
